Move all logic for newPerson to mixin
