+++ src/mixins/GrampsjsNewPersonMixin.js
@@ -7,6 +7,11 @@ import '../components/GrampsjsFormSelectObjectList.js'
 import '../components/GrampsjsFormSelectType.js'
 import '../components/GrampsjsFormPrivate.js'
 
+import {makeHandle, dateIsEmpty} from '../util.js'
+import {apiPost} from '../api.js'
+
+const dataDefault = {_class: 'Person', gender: 2, citation_list: []}
+
 const gender = {
   2: 'Unknown',
   1: 'Male',
@@ -15,6 +20,14 @@ const gender = {
 
 export const GrampsjsNewPersonMixin = superClass =>
   class extends superClass {
+    constructor() {
+      super()
+      this.data = dataDefault
+      this.postUrl = '/api/objects/'
+      this.itemPath = 'person'
+      this.objClass = 'Person'
+    }
+
     handleGender(e) {
       this.data = {...this.data, gender: parseInt(e.target.value, 10)}
     }
@@ -81,5 +94,152 @@ export const GrampsjsNewPersonMixin = superClass =>
           .strings="${this.strings}"
         ></grampsjs-form-private>
       `
+    }
+
+    _handleFormData(e) {
+      super._handleFormData(e)
+      const originalTarget = e.composedPath()[0]
+      if (originalTarget.id === 'primary-name') {
+        this.data = {...this.data, primary_name: e.detail.data}
+      }
+      if (originalTarget.id === 'birth-date') {
+        this.data = {
+          ...this.data,
+          birth: {...this.data.birth, date: e.detail.data},
+        }
+      }
+      if (originalTarget.id === 'birth-place-list') {
+        this.data = {
+          ...this.data,
+          birth: {...this.data.birth, place: e.detail.data[0]},
+        }
+      }
+      if (originalTarget.id === 'death-date') {
+        this.data = {
+          ...this.data,
+          death: {...this.data.death, date: e.detail.data},
+        }
+      }
+      if (originalTarget.id === 'death-place-list') {
+        this.data = {
+          ...this.data,
+          death: {...this.data.death, place: e.detail.data[0]},
+        }
+      }
+      this.checkFormValidity()
+    }
+
+    _processedData() {
+      const handlePerson = makeHandle()
+      const handleBirth = makeHandle()
+      const handleDeath = makeHandle()
+      const birthString = this.translateTypeName(false, 'event_types', 'Birth')
+      const deathString = this.translateTypeName(false, 'event_types', 'Death')
+      const {birth, death, ...person} = this.data
+      const hasBirth = birth.place || (birth?.date && !dateIsEmpty(birth.date))
+      const hasDeath = death.place || (death?.date && !dateIsEmpty(birth.date))
+      if (!hasBirth && !hasDeath) {
+        return [person]
+      }
+      if (!hasDeath) {
+        return [
+          {
+            ...person,
+            handle: handlePerson,
+            event_ref_list: [{_class: 'EventRef', ref: handleBirth}],
+            birth_ref_index: 0,
+          },
+          {
+            ...birth,
+            _class: 'Event',
+            handle: handleBirth,
+            type: {_class: 'EventType', string: birthString},
+            citation_list: person.citation_list ?? [],
+          },
+        ]
+      }
+      if (!hasBirth) {
+        return [
+          {
+            ...person,
+            handle: handlePerson,
+            event_ref_list: [{_class: 'EventRef', ref: handleDeath}],
+            death_ref_index: 0,
+          },
+          {
+            ...death,
+            _class: 'Event',
+            handle: handleDeath,
+            type: {_class: 'EventType', string: deathString},
+            citation_list: person.citation_list ?? [],
+          },
+        ]
+      }
+      return [
+        {
+          ...person,
+          handle: handlePerson,
+          event_ref_list: [
+            {_class: 'EventRef', ref: handleBirth},
+            {_class: 'EventRef', ref: handleDeath},
+          ],
+          birth_ref_index: 0,
+          death_ref_index: 1,
+        },
+        {
+          ...birth,
+          _class: 'Event',
+          handle: handleBirth,
+          type: {_class: 'EventType', string: birthString},
+          citation_list: person.citation_list ?? [],
+        },
+        {
+          ...death,
+          _class: 'Event',
+          handle: handleDeath,
+          type: {_class: 'EventType', string: deathString},
+          citation_list: person.citation_list ?? [],
+        },
+      ]
+    }
+
+    _submit() {
+      const processedData = this._processedData()
+      apiPost(this.postUrl, processedData).then(data => {
+        if ('data' in data) {
+          this.error = false
+          const grampsId = data.data.filter(
+            obj => obj.new._class === 'Person'
+          )[0].new.gramps_id
+          this.dispatchEvent(
+            new CustomEvent('nav', {
+              bubbles: true,
+              composed: true,
+              detail: {path: this._getItemPath(grampsId)},
+            })
+          )
+          this._reset()
+        } else if ('error' in data) {
+          this.error = true
+          this._errorMessage = data.error
+        }
+      })
+    }
+
+    checkFormValidity() {
+      let valid = true
+      this.shadowRoot
+        .querySelectorAll('grampsjs-form-select-date')
+        .forEach(element => {
+          if (!element.isValid()) {
+            valid = false
+          }
+        })
+      this.isFormValid = valid
+    }
+
+    _reset() {
+      super._reset()
+      this.data = dataDefault
     }
   }
